requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
blis==0.7.11
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
distro==1.9.0
fastapi==0.104.1
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl#sha256=37e9f1f6a278a5138fabdabcc92cc559da917f9b24c76f0adf6758720d7eab10
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
langcodes==3.5.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numpy==1.26.4
openai==1.3.0
packaging==26.0
pgvector==0.2.4
preshed==3.0.12
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.0.0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.1
setuptools==80.10.2
shellingham==1.5.4
smart_open==7.5.0
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.5.2
starlette==0.27.0
thinc==8.2.5
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.24.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==16.0
wrapt==2.0.1
